Santali Dictionary
The Santali Dictionary is an open-source project aimed at preserving and promoting the Santali language by providing a comprehensive and accessible digital dictionary. This project allows users to look up Santali words, their meanings, and related information. It serves as a resource for both native speakers and those interested in learning the language.

The dictionary is organized to include not just translations but also cultural context and usage examples, making it a valuable tool for education and linguistic research.

Developer:
Kailash Murmu